+++ internal/githubauth/types.go
@@ -17,7 +17,11 @@ const (
 	MaxRequestBytes = 64 << 10
 
 	ErrorCodeLocalPassphraseRequired = "local_passphrase_required"
-	MinGrantDuration                 = 15 * time.Minute
+	ApprovalTimeout                  = 15 * time.Minute
+	PostApprovalExchangeTimeout      = 2 * time.Minute
+	BrokerExchangeTimeout            = ApprovalTimeout + PostApprovalExchangeTimeout
+	MinimumUsableGrantDuration       = 15 * time.Minute
+	MinGrantDuration                 = ApprovalTimeout + MinimumUsableGrantDuration
 	AbsoluteMaxGrantDuration         = 24 * time.Hour
 )